Working neural network with 99% accuracy
